requirements.txt
pyTelegramBotAPI==3.8.2
Pillow==8.3.1
